Change get stock card by id to query by facility as well
Change API endpoint of getting stock card by id, to get by facilityid and id. This is because of how the API URL is constructed. The client gets a stock card by id, but also specifies a facility id. Do not ignore facility id, as if we just check by stock card id, the client might (incorrectly) assume that stock card is present at that facility. If a stock card id is not present at that facility, just say the stock card does not exist, even though it exists, just not at that facility.