VIMS-381 Add lots and lots on hand to adjustment API
Adjustment API needs support for lots, to update lots and their on hand quantity
. This adds support for that, including creating new lots if lot information is
specified and lot is not found in the system. If only lot id is specified and no
t found in the system, gives an HTTP Bad Request error.

Adds support for alternative way to define lot besides lot id in stock event DTO
.