VIMS-381 Look for lots on hand based on lot AND stock card
Change method calls to backend to get lots on hand by lot AND stock card. This is because lots on hand is unique based on both information. Add code to update lot on hand information if existing.

Fix various bugs in the mapper code. Change update methods in lot and stock card mappers to update effective and modified dates in db when update has been performed.