VIMS-389 Add customProps attribute to lotsOnHand
keyValues/customProps was removed in previous commit (da872da) because it was returning too many entries in database. This is the fix. Fill in keyValues data from database, but do not return it in the JSON. Instead, a customProps attribute is provided, and the getCustomProps method fills in this attribute, from keyValues.

There are two classes where keyValues is hidden and customProps is used instead: StockCardEntry and LotsOnHand. In StockCardEntry, the getCustomProps simply transforms the keyValues list into a map and returns that in JSON format. In LotsOnHand, an aggregation is done. This is a multi-step process:

* Get all the keys in keyValues
* Remove duplicates
* Go through each key, get all its values and choose one value, using a strategy (StockCardEntryKVReduceStrategy interface)
* Add the key and value to the map
* Return the map as customProps

The default strategy is to choose the value that has the latest recordedDate (MaxRecordedStrategy), but another class can implement StockCardEntryKVReduceStrategy.

This commit also adds recordedDate to StockCardEntryKV, and adds mapper code to support setting keyValues in both objects.