VIMS-398 Add permissions checks to stock card API calls
Add permissions checks to the three GET and one POST stock card API calls.

For the GET calls, before the call is made to the service, a list of rights is retrieved for the user, by facility and product. If the view stock on hand right is found in the list, then the service call is made, otherwise a 403 Forbidden error is returned. For the API call that gets multiple stock cards, if no stock cards exist, then a 404 Not Found is returned--otherwise, an empty stock card list is returned.

For the POST call, each stock event is checked, for a list of rights by user, facility and product. Each stock event needs to be checked, because different events might specify different products. The check returns a list of rights, and manage stock permission is searched in the list. If not found, then a 403 Forbidden error is returned. This means if any event has incorrect permissions, the whole API call does not go through.

Also, add test cases for the new code, and add Swagger API documentation for the changes.